Fixed some redirection and heredoc issues

+++ srcs/history.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   history.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: bleow <bleow@student.42kl.edu.my>          +#+  +:+       +#+        */
+/*   By: lechan <lechan@student.42kl.edu.my>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/01 06:29:46 by bleow             #+#    #+#             */
-/*   Updated: 2025/03/19 23:39:40 by bleow            ###   ########.fr       */
+/*   Updated: 2025/03/22 19:52:30 by lechan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,21 +28,21 @@ Example: init_history_fd(O_RDONLY)
 */
 int	init_history_fd(int mode)
 {
-    int	fd;
+	int	fd;
 
-    if (access(HISTORY_FILE, F_OK) == -1)
-    {
-        fd = open(HISTORY_FILE, O_WRONLY | O_CREAT, 0644);
-        if (fd == -1)
-            return (-1);
-        close(fd);
-    }
-    if (mode == O_RDONLY && access(HISTORY_FILE, R_OK) == -1)
-        return (-1);
-    if (mode == O_WRONLY && access(HISTORY_FILE, W_OK) == -1)
-        return (-1);
-    fd = open(HISTORY_FILE, mode);
-    return (fd);
+	if (access(HISTORY_FILE, F_OK) == -1)
+	{
+		fd = open(HISTORY_FILE, O_WRONLY | O_CREAT, 0644);
+		if (fd == -1)
+			return (-1);
+		close(fd);
+	}
+	if (mode == O_RDONLY && access(HISTORY_FILE, R_OK) == -1)
+		return (-1);
+	if (mode == O_WRONLY && access(HISTORY_FILE, W_OK) == -1)
+		return (-1);
+	fd = open(HISTORY_FILE, mode);
+	return (fd);
 }
 
 /*
@@ -60,18 +60,18 @@ Example: append_history(fd, "ls -la")
 */
 int	append_history(int fd, const char *line)
 {
-    ssize_t	write_ret;
+	int	write_ret;
 
-    if (!line)
-        return (0);
-    write_ret = write(fd, line, ft_strlen(line));
-    if (write_ret == -1)
-        return (0);
-    write_ret = write(fd, "\n", 1);
-    if (write_ret == -1)
-        return (0);
-    add_history(line);
-    return (1);
+	if (!line)
+		return (0);
+	write_ret = write(fd, line, ft_strlen(line));
+	if (write_ret == -1)
+		return (0);
+	write_ret = write(fd, "\n", 1);
+	if (write_ret == -1)
+		return (0);
+	add_history(line);
+	return (1);
 }
 
 /*
@@ -90,21 +90,21 @@ Example: If history file contains 100 commands
 */
 int	get_history_count(void)
 {
-    int		fd;
-    int		count;
-    char	*line;
+	int		fd;
+	int		count;
+	char	*line;
 
-    count = 0;
-    fd = init_history_fd(O_RDONLY);
-    if (fd == -1)
-        return (0);
-    line = get_next_line(fd);
-    while (line)
-    {
-        count++;
-        ft_safefree((void **)&line);
-        line = get_next_line(fd);
-    }
-    close(fd);
-    return (count);
+	count = 0;
+	fd = init_history_fd(O_RDONLY);
+	if (fd == -1)
+		return (0);
+	line = get_next_line(fd);
+	while (line)
+	{
+		count++;
+		ft_safefree((void **)&line);
+		line = get_next_line(fd);
+	}
+	close(fd);
+	return (count);
 }
